♻️ refactor #68: atomic design에 맞춰 디렉토리 구조 변경
