--- fe/src/components/common/Icon/IconComponents.tsx
@@ -1,39 +1,0 @@
-import { ReactComponent as account } from '@assets/account.svg';
-import { ReactComponent as arrowUp } from '@assets/arrowUp.svg';
-import { ReactComponent as camera } from '@assets/camera.svg';
-import { ReactComponent as cancel } from '@assets/cancel.svg';
-import { ReactComponent as category } from '@assets/category.svg';
-import { ReactComponent as chevronDown } from '@assets/chevronDown.svg';
-import { ReactComponent as chevronLeft } from '@assets/chevronLeft.svg';
-import { ReactComponent as home } from '@assets/home.svg';
-import { ReactComponent as keyboard } from '@assets/keyboard.svg';
-import { ReactComponent as like } from '@assets/like.svg';
-import { ReactComponent as message } from '@assets/message.svg';
-import { ReactComponent as more } from '@assets/more.svg';
-import { ReactComponent as orderCircle } from '@assets/orderCircle.svg';
-import { ReactComponent as plus } from '@assets/plus.svg';
-import { ReactComponent as search } from '@assets/search.svg';
-import { ReactComponent as soldList } from '@assets/soldList.svg';
-import { ReactComponent as townSetting } from '@assets/townSetting.svg';
-
-const IconComponents = {
-  account,
-  arrowUp,
-  camera,
-  cancel,
-  category,
-  chevronDown,
-  chevronLeft,
-  home,
-  keyboard,
-  like,
-  message,
-  more,
-  orderCircle,
-  plus,
-  search,
-  soldList,
-  townSetting,
-};
-
-export { IconComponents };
