--- fe/src/components/common/Icon/custom.d.ts
@@ -1,6 +1,0 @@
-declare module '*.svg' {
-  import React = require('react');
-  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
-  const src: string;
-  export default src;
-}
