+++ fe/src/components/molecules/TabBars/DetailTabBar/DetailTabBar.stories.tsx
