+++ fe/src/components/molecules/TabBars/DetailTabBar/DetailTabBar.style.ts
