+++ fe/src/components/molecules/TabBars/SaleTabBar/SaleTabBar.style.ts
